# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "ATWT API for Testing",
        "description": "This is an example api for testing purpose. automatedtestingwithtuyen.com",
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "contact": {
            "name": "Tuyen Nguyen",
            "url": "http://automatedtestingwithtuyen.com",
            "email": "trongtuyen96@gmail.com"
        }
    },
    "host": "atwt-api.herokuapp.com",
    "basePath": "/v1",
...
    "schemes": [
        "https"
    ],
    "paths": {
...
        "/factory": {
            "get": {
                "tags": [
                    "factory"
                ],
                "summary": "Get factories with conditions",
                "description": "",
                "operationId": "getFactories",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "name",
                        "description": "Name of factory",
                        "type": "string"
                    },
                    {
                        "in": "query",
                        "name": "country",
                        "description": "Country of factory",
                        "type": "string"
                    },
                    {
                        "in": "query",
                        "name": "license",
                        "description": "License of factory",
                        "type": "string"
                    },
                    {
                        "in": "query",
                        "name": "year",
                        "description": "Year of establishment of factory (greater than or equal)",
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfull operation",
                        "schema": {
                            "type": "array",
                            "$ref": "#/definitions/Factory"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "factory"
                ],
                "summary": "Add a new factory",
                "description": "",
                "operationId": "postFactory",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Factory object that need to add",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Factory"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfull operation"
                    },
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/factory/{factoryId}": {
            "get": {
                "tags": [
                    "factory"
                ],
                "summary": "Find factory by Id",
                "description": "Returns a single factory",
                "operationId": "getFactoryById",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "factoryId",
                        "in": "path",
                        "description": "Id of factory to return",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/Factory"
                        }
                    },
                    "400": {
                        "description": "Invalid Id supplied"
                    },
                    "404": {
                        "description": "Factory not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "factory"
                ],
                "summary": "Update an existing factory",
                "description": "",
                "operationId": "putFactory",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "factoryId",
                        "in": "path",
                        "description": "Id of factory to update",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Factory object that needs to be updated",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Factory"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Update successfully"
                    },
                    "400": {
                        "description": "Invalid Id supplied"
                    },
                    "404": {
                        "description": "Factory not found"
                    },
                    "405": {
                        "description": "Validation exception"
                    }
                }
            },
            "delete": {
                "tags": [
                    "factory"
                ],
                "summary": "Delete a factory",
                "description": "",
                "operationId": "deleteFactory",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "factoryId",
                        "in": "path",
                        "description": "Factory Id to delete",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid Id supplied"
                    },
                    "404": {
                        "description": "Factory not found"
                    }
                },
                "deprecated": true
            }
        },
        "/factory/{factoryId}/foods": {
            "get": {
                "tags": [
                    "factory"
                ],
                "summary": "Find all foods produced by factory by Id",
                "description": "Returns foods produced by factory",
                "operationId": "getAllFoodsOfFactoryById",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "factoryId",
                        "in": "path",
                        "description": "Id of factory to return",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "type": "array",
                            "$ref": "#/definitions/Food"
                        }
                    },
                    "400": {
                        "description": "Invalid Id supplied"
                    },
                    "404": {
                        "description": "Factory not found"
                    }
                }
            }
        },
        "/food": {
            "get": {
                "tags": [
                    "food"
                ],
                "summary": "Get foods with conditions",
                "description": "",
                "operationId": "getFoods",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "name",
                        "description": "Name of food",
                        "type": "string"
                    },
                    {
                        "in": "query",
                        "name": "color",
                        "description": "Color of food",
                        "type": "string"
                    },
                    {
                        "in": "query",
                        "name": "origin",
                        "description": "Origin of food",
                        "type": "string"
                    },
                    {
                        "in": "query",
                        "name": "weight",
                        "description": "Weight of food (greater than or equal)",
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfull operation",
                        "schema": {
                            "type": "array",
                            "$ref": "#/definitions/Food"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "food"
                ],
                "summary": "Add a new food",
                "description": "",
                "operationId": "postFood",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Food object that need to add",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Food"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfull operation"
                    },
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/food/{foodId}": {
            "get": {
                "tags": [
                    "food"
                ],
                "summary": "Find food by Id",
                "description": "Returns a single food",
                "operationId": "getFoodById",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "foodId",
                        "in": "path",
                        "description": "Id of food to return",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/Food"
                        }
                    },
                    "400": {
                        "description": "Invalid Id supplied"
                    },
                    "404": {
                        "description": "Food not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "food"
                ],
                "summary": "Update an existing food",
                "description": "",
                "operationId": "putFood",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "foodId",
                        "in": "path",
                        "description": "Id of food to update",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Food object that needs to be updated",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Food"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Update successfully"
                    },
                    "400": {
                        "description": "Invalid Id supplied"
                    },
                    "404": {
                        "description": "Food not found"
                    },
                    "405": {
                        "description": "Validation exception"
                    }
                }
            },
            "delete": {
                "tags": [
                    "food"
                ],
                "summary": "Delete a food",
                "description": "",
                "operationId": "deleteFood",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "foodId",
                        "in": "path",
                        "description": "Food Id to delete",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid Id supplied"
                    },
                    "404": {
                        "description": "Food not found"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Factory": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "license": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            },
            "xml": {
                "name": "Factory"
            }
        },
        "Food": {
            "type": "object",
            "required": [
                "factoryID"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "weight": {
                    "type": "integer"
                },
                "color": {
                    "type": "string"
                },
                "origin": {
                    "type": "string"
                },
                "factoryID": {
                    "type": "string"
                }
            },
            "xml": {
                "name": "Food"
            }
        },
...
    }
}